services/detection-service/requirements.txt
redis
fastapi
uvicorn
python-dotenv
psycopg[binary]
scikit-learn
pandas
numpy
prometheus-client
joblib